+++ ui-lib/src/main/java/co/electriccoin/zcash/ui/screen/history/HistoryTag.kt
@@ -5,5 +5,6 @@ package co.electriccoin.zcash.ui.screen.history
  */
 object HistoryTag {
     const val TRANSACTION_LIST = "transaction_list"
+    const val TRANSACTION_ITEM = "transaction_item"
     const val PROGRESS = "progress_bar"
 }