+++ ui-lib/src/main/java/co/electriccoin/zcash/ui/screen/request/AndroidRequest.kt
@@ -30,7 +30,9 @@ private fun WrapRequest(
     val walletAddresses = walletViewModel.addresses.collectAsStateWithLifecycle().value
 
     if (null == walletAddresses) {
-        // Improve this by allowing screen composition and updating it after the data is available
+        // TODO [#1146]: Consider moving CircularScreenProgressIndicator from Android layer to View layer
+        // TODO [#1146]: Improve this by allowing screen composition and updating it after the data is available
+        // TODO [#1146]: https://github.com/Electric-Coin-Company/zashi-android/issues/1146
         CircularScreenProgressIndicator()
     } else {
         Request(