+++ ui-lib/src/main/java/co/electriccoin/zcash/ui/screen/scan/AndroidScan.kt
@@ -41,7 +41,9 @@ fun WrapScan(
     val scope = rememberCoroutineScope()
 
     if (synchronizer == null) {
-        // Improve this by allowing screen composition and updating it after the data is available
+        // TODO [#1146]: Consider moving CircularScreenProgressIndicator from Android layer to View layer
+        // TODO [#1146]: Improve this by allowing screen composition and updating it after the data is available
+        // TODO [#1146]: https://github.com/Electric-Coin-Company/zashi-android/issues/1146
         CircularScreenProgressIndicator()
     } else {
         Scan(