+++ ui-lib/src/main/java/co/electriccoin/zcash/ui/screen/seedrecovery/AndroidSeedRecovery.kt
@@ -40,7 +40,9 @@ private fun WrapSeedRecovery(
     val synchronizer = walletViewModel.synchronizer.collectAsStateWithLifecycle().value
 
     if (null == synchronizer || null == persistableWallet) {
-        // Improve this by allowing screen composition and updating it after the data is available
+        // TODO [#1146]: Consider moving CircularScreenProgressIndicator from Android layer to View layer
+        // TODO [#1146]: Improve this by allowing screen composition and updating it after the data is available
+        // TODO [#1146]: https://github.com/Electric-Coin-Company/zashi-android/issues/1146
         CircularScreenProgressIndicator()
     } else {
         SeedRecovery(